场内扫码查询订单信息 停车场微信扫码下单 停车场微信扫码下单回调 通道码扫码查询订单

+++ src/main/java/com/example/jiuzhou/user/model/ParkPayRecord.java
@@ -6,6 +6,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.Column;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -26,7 +27,7 @@ public class ParkPayRecord {
     private Integer TenantId;
 
     @Column(name="Money")
-    private Integer Money;
+    private BigDecimal Money;
 
     @Column(name="ParkadeAccessId")
     private Integer ParkadeAccessId;
